backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.20.2
gunicorn==21.2.0
requests-oauthlib==1.3.1
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0